Factor our notification management in IMPORTER
