+++ libvast/vast/system/importer.hpp
@@ -104,6 +104,9 @@ struct importer_state {
   /// @returns various status metrics.
   caf::dictionary<caf::config_value> status() const;
 
+  /// Sends a notification to all listeners and clears the listeners vector.
+  void notify_flush_listeners();
+
   /// Stores how many slices inbound paths can still send us.
   int32_t in_flight_slices = 0;
